+++ Tunny/Solver/Optuna.cs
@@ -10,6 +10,7 @@ using Python.Runtime;
 
 using Tunny.Handler;
 using Tunny.Settings;
+using Tunny.Type;
 using Tunny.UI;
 using Tunny.Util;
 
@@ -34,6 +35,7 @@ namespace Tunny.Solver
         public bool RunSolver(
             List<Variable> variables,
             IEnumerable<IGH_Param> objectives,
+            Dictionary<string, FishEgg> fishEggs,
             Func<ProgressState, int, EvaluatedGHResult> evaluate)
         {
             string[] objNickName = objectives.Select(x => x.NickName).ToArray();
@@ -45,7 +47,7 @@ namespace Tunny.Solver
 
             try
             {
-                var optimize = new Algorithm(variables, _hasConstraint, objNickName, _settings, Eval);
+                var optimize = new Algorithm(variables, _hasConstraint, objNickName, fishEggs, _settings, Eval);
                 optimize.Solve();
                 XOpt = optimize.GetXOptimum();