+++ Tunny/Util/GrasshopperInOut.cs
@@ -21,7 +21,7 @@ namespace Tunny.Util
     {
         private readonly GH_Document _document;
         private readonly List<Guid> _inputGuids;
-        private readonly TunnyComponent _component;
+        private readonly GH_Component _component;
         private List<GalapagosGeneListObject> _genePool;
         private GH_FishAttribute _attributes;
 
@@ -29,24 +29,27 @@ namespace Tunny.Util
         public List<GH_NumberSlider> Sliders { get; set; }
         public string ComponentFolder { get; }
         public List<Variable> Variables { get; set; }
+        public Dictionary<string, FishEgg> EnqueueItems { get; set; }
         public bool HasConstraint { get; set; }
         public string DocumentPath { get; set; }
         public string DocumentName { get; set; }
 
-        public GrasshopperInOut(TunnyComponent component)
+        public GrasshopperInOut(GH_Component component, bool getVariableOnly = false)
         {
             _component = component;
             ComponentFolder = Path.GetDirectoryName(Grasshopper.Instances.ComponentServer.FindAssemblyByObject(_component).Location);
             _document = _component.OnPingDocument();
             _inputGuids = new List<Guid>();
-            SetInputs();
-        }
-
-        private void SetInputs()
-        {
-            SetVariables();
-            SetObjectives();
-            SetAttributes();
+            if (getVariableOnly)
+            {
+                SetVariables();
+            }
+            else
+            {
+                SetVariables();
+                SetObjectives();
+                SetAttributes();
+            }
         }
 
         private void SetVariables()
@@ -62,16 +65,16 @@ namespace Tunny.Util
             }
 
             var variables = new List<Variable>();
-            SortInputs();
+            FilterInputVariables();
             SetInputSliderValues(variables);
             SetInputGenePoolValues(variables);
             Variables = variables;
         }
 
-        private void SortInputs()
+        private void FilterInputVariables()
         {
             var errorGuids = new List<Guid>();
-            foreach ((IGH_DocumentObject docObject, int i) in _inputGuids.Select((guid, i) => (_document.FindObject(guid, true), i)))
+            foreach ((IGH_DocumentObject docObject, int i) in _inputGuids.Select((guid, i) => (_document.FindObject(guid, false), i)))
             {
                 switch (docObject)
                 {
@@ -81,11 +84,20 @@ namespace Tunny.Util
                     case GalapagosGeneListObject genePool:
                         _genePool.Add(genePool);
                         break;
+                    case Param_FishEgg fishEgg:
+                        var ghFishEgg = (GH_FishEgg)fishEgg.VolatileData.AllData(true).First();
+                        EnqueueItems = ghFishEgg.Value;
+                        break;
                     default:
                         errorGuids.Add(docObject.InstanceGuid);
                         break;
                 }
             }
+            CheckHasIncorrectVariableInput(errorGuids);
+        }
+
+        private void CheckHasIncorrectVariableInput(List<Guid> errorGuids)
+        {
             if (errorGuids.Count > 0)
             {
                 ShowIncorrectVariableInputMessage(errorGuids);
@@ -110,6 +122,8 @@ namespace Tunny.Util
             {
                 decimal min = slider.Slider.Minimum;
                 decimal max = slider.Slider.Maximum;
+                decimal value = slider.Slider.Value;
+                Guid id = slider.InstanceGuid;
 
                 decimal lowerBond;
                 decimal upperBond;
@@ -144,7 +158,7 @@ namespace Tunny.Util
                         break;
                 }
 
-                variables.Add(new Variable(Convert.ToDouble(lowerBond), Convert.ToDouble(upperBond), isInteger, nickName, eps));
+                variables.Add(new Variable(Convert.ToDouble(lowerBond), Convert.ToDouble(upperBond), isInteger, nickName, eps, Convert.ToDouble(value), id));
             }
         }
 
@@ -160,10 +174,13 @@ namespace Tunny.Util
                 decimal lowerBond = genePool.Minimum;
                 decimal upperBond = genePool.Maximum;
                 double eps = Math.Pow(10, -genePool.Decimals);
+                Guid id = genePool.InstanceGuid;
                 for (int j = 0; j < genePool.Count; j++)
                 {
+                    IGH_Goo[] goo = genePool.VolatileData.AllData(false).ToArray();
+                    var ghNumber = (GH_Number)goo[j];
                     string name = nickNames[i] + j;
-                    variables.Add(new Variable(Convert.ToDouble(lowerBond), Convert.ToDouble(upperBond), isInteger, name, eps));
+                    variables.Add(new Variable(Convert.ToDouble(lowerBond), Convert.ToDouble(upperBond), isInteger, name, eps, ghNumber.Value, id));
                 }
             }
         }